changed return type of the inserted user from long to int

+++ app/src/main/java/tmidev/localaccount/domain/usecase/InsertUserUseCase.kt
@@ -5,12 +5,12 @@ import tmidev.localaccount.domain.model.User
 import javax.inject.Inject
 
 interface InsertUserUseCase {
-    suspend operator fun invoke(user: User): Long
+    suspend operator fun invoke(user: User): Int
 }
 
 class InsertUserUseCaseImpl @Inject constructor(
     private val usersRepository: UsersRepository
 ) : InsertUserUseCase {
-    override suspend fun invoke(user: User): Long =
+    override suspend fun invoke(user: User): Int =
         usersRepository.insertUser(user = user)
 }
